fix overflow probrem of Timer
changed unsigned int -> unsigned long int

+++ src/common/Timer.cpp
@@ -11,7 +11,7 @@ using namespace std;
 Timer::Timer(){
   restart();
 }
-    
+
 Timer::~Timer(){}
 
 void Timer::restart() {
@@ -33,22 +33,22 @@ chrono::nanoseconds Timer::get_time() const{
   return chrono::duration_cast<chrono::nanoseconds>(end_point-start_point_);
 }
 
-unsigned int Timer::get_elapsed_h() const{
+unsigned long int Timer::get_elapsed_h() const{
   return chrono::duration_cast<chrono::hours>(get_time()).count();
 }
-unsigned int Timer::get_elapsed_m() const{
+unsigned long int Timer::get_elapsed_m() const{
   return chrono::duration_cast<chrono::minutes>(get_time()).count();
 }
-unsigned int Timer::get_elapsed_s() const{
+unsigned long int Timer::get_elapsed_s() const{
   return chrono::duration_cast<chrono::seconds>(get_time()).count();
 }
-unsigned int Timer::get_elapsed_ms() const{
+unsigned long int Timer::get_elapsed_ms() const{
   return chrono::duration_cast<chrono::milliseconds>(get_time()).count();
 }
-unsigned int Timer::get_elapsed_us() const{
+unsigned long int Timer::get_elapsed_us() const{
   return chrono::duration_cast<chrono::microseconds>(get_time()).count();
 }
-unsigned int Timer::get_elapsed_ns() const{
+unsigned long int Timer::get_elapsed_ns() const{
   return get_time().count();
 }
